refactored mock apis to allow for mocking backend
