+++ test/source/mock/google/strategies/strategy-base.ts
