--- test/source/mock/types.ts
@@ -1,6 +1,0 @@
-export type DraftSaveModel = {
-    message: {
-        raw: string,
-        threadId: string
-    }
-};
